requirements.txt
fastapi
uvicorn
langchain==0.3.26
langchain-core==0.3.67
langchain-community==0.3.26
langchain-huggingface==0.3.0
openai
langgraph
psycopg2-binary
sqlalchemy
python-dotenv
yfinance
requests
cachetools
wikipedia
pypdf
sentence-transformers
chromadb==0.4.24
pydantic-settings  
numpy<2
torch==2.2.2

